+++ src/Normalizer/StatementNormalizer.php
@@ -11,6 +11,7 @@
 
 namespace Xabbuh\XApi\Serializer\Symfony\Normalizer;
 
+use Xabbuh\XApi\Common\Exception\UnsupportedStatementVersionException;
 use Xabbuh\XApi\Model\Statement;
 use Xabbuh\XApi\Model\StatementId;
 
@@ -64,6 +65,10 @@ final class StatementNormalizer extends Normalizer
             $data['attachments'] = $this->normalizeAttribute($attachments, $format, $context);
         }
 
+        if (null !== $version = $object->getVersion()) {
+            $data['version'] = $version;
+        }
+
         return $data;
     }
 
@@ -80,6 +85,16 @@ final class StatementNormalizer extends Normalizer
      */
     public function denormalize($data, $class, $format = null, array $context = array())
     {
+        $version = null;
+
+        if (isset($data['version'])) {
+            $version = $data['version'];
+
+            if (!preg_match('/^1\.0(?:\.\d+)?$/', $version)) {
+                throw new UnsupportedStatementVersionException(sprintf('Statements at version "%s" are not supported.', $version));
+            }
+        }
+
         $id = isset($data['id']) ? StatementId::fromString($data['id']) : null;
         $actor = $this->denormalizeData($data['actor'], 'Xabbuh\XApi\Model\Actor', $format, $context);
         $verb = $this->denormalizeData($data['verb'], 'Xabbuh\XApi\Model\Verb', $format, $context);
@@ -115,7 +130,7 @@ final class StatementNormalizer extends Normalizer
             $attachments = $this->denormalizeData($data['attachments'], 'Xabbuh\XApi\Model\Attachment[]', $format, $context);
         }
 
-        return new Statement($id, $actor, $verb, $object, $result, $authority, $created, $stored, $statementContext, $attachments);
+        return new Statement($id, $actor, $verb, $object, $result, $authority, $created, $stored, $statementContext, $attachments, $version);
     }
 
     /**